Move some packets outside
Mostly because these packets are used in configuration phase now too
