+++ Obsidian/Net/Packets/Configuration/Clientbound/UpdateTagsPacket.cs
@@ -8,19 +8,10 @@ public sealed partial class UpdateTagsPacket : IClientboundPacket
 
     public int Id => 0x08;
 
-    public static UpdateTagsPacket FromRegistry { get; } = new(TagsRegistry.Categories);
+    public static UpdateTagsPacket FromRegistry { get; } = new(Registries.TagsRegistry.Categories);
 
     public UpdateTagsPacket(IDictionary<string, Tag[]> tags)
     {
         this.Tags = tags;
     }
-}
-
-public class Tag
-{
-    public string Name { get; init; }
-    public string Type { get; init; }
-    public bool Replace { get; init; }
-    public int[] Entries { get; init; }
-    public int Count => Entries.Length;
 }
