+++ Obsidian/Net/Packets/Configuration/Serverbound/FinishConfigurationPacket.cs
@@ -5,7 +5,7 @@ public sealed partial class FinishConfigurationPacket : IServerboundPacket
 {
     public int Id => 0x02;
 
-    public ValueTask HandleAsync(Server server, Player player) => ValueTask.CompletedTask
+    public ValueTask HandleAsync(Server server, Player player) => ValueTask.CompletedTask;
     public void Populate(byte[] data) { }
     public void Populate(MinecraftStream stream) { }
 }
