+++ Obsidian/Net/Packets/DisconnectPacket.cs
@@ -11,7 +11,7 @@ public partial class DisconnectPacket : IClientboundPacket
 
     public DisconnectPacket(ChatMessage reason, ClientState state)
     {
-        Id = state == ClientState.Play ? 0x1A : 0x00;
+        Id = state == ClientState.Configuration ? 0x01 : state == ClientState.Play ? 0x1A : 0x00;
         Reason = reason;
     }
 }
