+++ Obsidian/Net/Packets/KeepAlivePacket.cs
@@ -7,10 +7,11 @@ public partial class KeepAlivePacket : IClientboundPacket, IServerboundPacket
     [Field(0)]
     public long KeepAliveId { get; private set; }
 
-    public int Id => 0x23;
+    public int Id { get; init; }
 
     public KeepAlivePacket()
     {
+
     }
 
     public KeepAlivePacket(long id)
